libsubprocess: relocate buffer code
Problem: the flux buffer classes are part of the Flux API but are
neither widely used nor documented.

Since the primary user is libsubprocess, relocate the flux buffer classes
and their tests to libsubprocess.

The only other changes in this commit are those necessary to get
everything to compile again.
